Big refactoring to move from state oriented code
The refactorings in this commit are a work in progress towards a code
which is composed mostly of pure functions. The trick here is to stop
using locations and instead use places all over. I added a location to
each place in the world, thus making the location (x, y) less important
and reducing state changes to a single call to world/update-place.